Change to the contacts section config format
I wanted to add a link for a static pdf resume or similar, and found
that the font awesome pdf file icon used a different prefix to the one
that was already hard-coded in this template. So I changed the contacts
section config format to be more flexible to accommodate this.